Code-split Svelte components
