+++ package.json
@@ -47,6 +47,7 @@
     "set-json-version": "^1.0.6",
     "svelte": "^3.43.2",
     "svelte-materialify": "0.3.7",
+    "svelte-preprocess": "^4.9.8",
     "tslib": "^2.3.1",
     "typescript": "^4.5.2",
     "web-ext": "^6.5.0",
